requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiosqlite==0.19.0
python-dotenv==1.0.0
Pillow==10.2.0
openai==1.12.0
geopy==2.4.1
asyncio==3.4.3
aiohttp==3.9.1
